chore: rename utils file for styling
