+++ frontend/src/components/retroui/Card.tsx
@@ -1,6 +1,6 @@
 import { type HTMLAttributes } from "react";
 
-import { cn } from "@/lib/tailwind.utils";
+import { cn } from "@/lib/retroui.utils";
 
 import { Text } from "./Text";
